Moved even more static utility methods out of main plugin class and into the utility class
